### 2021-08-31 #45 @enra64
Jointly: A Python Package for synchronizing multiple sensors with accelerometers
Submitting Author: Arne Herdick (@enra64)
Package Name: Jointly
One-Line Description of Package: Jointly is a python package for synchronizing sensors with accelerometer data
Repository Link:  https://github.com/hpi-dhc/jointly
Version submitted: 1.0.2
Editor: Ivan Ogasawara (@xmnlab)
Reviewer 1: Alejandro Sáez Mollejo (@AlexS12)
Reviewer 2: Arthur Endsley (@arthur-e)
Archive: https://zenodo.org/record/5770586#.Yd8AUoRByV5
Version accepted: 1.0.3

---

jointly is a python package for synchronizing devices that contain accelerometer sensors. For example, if you shake all your devices together before and after a study, jointly will find these shakes, remove any temporal offset between the sensors, and stretch the data so every clock aligns to a reference sensor. Jointly ingests and produces pandas' DataFrame objects.

\* Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [notes on categories][NotesOnCategories] of our guidebook.

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences):
This package was developed to synchronize multiple sensors together. Our usecase was to synchronize 6 IMUs to an ECG sensor with an IMU, all of which were worn by a participant during our study.

The target audience is fellow researchers trying to achieve high-quality synchronization for multiple study devices that also have an Accelerometer.

To our current knowledge, none exist.

- [x] has an [OSI approved license][OsiApprovedLicense].

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in maintaining my package should it be accepted.

[Editor and review templates can be found here][Templates]

[NotesOnCategories]: https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/aims-and-scope.html?highlight=data#notes-on-categories

[Templates]: https://www.pyopensci.org/contributing-guide/appendices/templates.html
