### 2021-11-03 #45 @xmnlab
@Alexs12 as we talked some days ago, we are very happy to have you here reviewing Jointly. thank you!

The PyOpenSci contributing guide is here https://www.pyopensci.org/contributing-guide/intro.html and this is the guide specifically for reviewers: https://www.pyopensci.org/contributing-guide/open-source-software-submissions/reviewer-guide.html

Feel free to ping me anytime when you need any information or help.

I am waiting for the confirmation of one more reviewer and after that, I will update the schedules for this review.

thank you so much for your support :)
