### 2021-11-03 #45 @xmnlab
@enra64 not sure if you were able to check my comment here: https://github.com/pyOpenSci/software-review/issues/45#issuecomment-950481738.

We often select one reviewer suggested by the package's author and another one will be selected by us. If you don't have any suggestions then I will look for another one. Thank you so much.
