### 2021-11-23 #45 @xmnlab
Hi @arthur-e! thank you so much for accepting to be a reviewer for the package Jointly!

I sent you some links but here there are two links you can check before you start.

- The PyOpenSci contributing guide: https://www.pyopensci.org/contributing-guide/intro.html
- The guide specifically for reviewers: https://www.pyopensci.org/contributing-guide/open-source-software-submissions/reviewer-guide.html

Feel free to ping me anytime when you need any information or help.

thank you so much for your support :)
