### 2021-11-23 #45 @xmnlab
@enra64 we have already two reviewers for Jointly, @AlexS12 and @arthur-e!

So, we are now starting the review process, and the deadline for this first stage is December, 13th.

In general, we aim for 3 weeks for review, 2 weeks for subsequent changes, and 1 week for reviewer approval of changes.

Let me know if you have any questions or comments, please.
