### 2021-12-08 #45 @arthur-e
- [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work.

The package *must be* improved with regards to the following criteria:

- [ ] **Function Documentation:** for all user-facing functions
  - [There is a page](https://jointly.readthedocs.io/en/latest/reference.html#jointly-synchronizer) that *lists* the functions but there is no documentation.
- [ ] **Examples** for all user-facing functions
  - Perhaps @xmnlab could elaborate here... I'm not sure if these examples are supposed to be in the API documentation or if a comprehensive "vignette" is sufficient.

The package *could be* improved with regards to the following criteria:

  - Both the "problems" and the "target audience" could be elaborated in the first paragraph of the README. Currently, this paragraph describes (adequately) what the package does, but does not describe the wider scientific and technical environment.
  - I was not able to install the package locally (in "editable" mode) due to a lack of support for `poetry-core>=1.2.0` in Python 3.8.12. The minimum Python version should probably be specified. As `jointly` could be installed from PyPI (`pip install jointly`), perhaps `poetry-core>=1.2.0` is *not* a hard requirement and could be changed.
  - There are "development version" instructions in `CONTRIBUTING.rst`, however.

  - Vignette from the README works great. However, I think that the plot that appears from `jointly.plot_reference_columns()` should have the axes labeled.
- [ ] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

#### README Requirements

  - (The README is an RST file.)

"The README should include, from top to bottom:"

The package *could* be improved with regards to:

- [x] **The package name**
  - I think that "jointly" would be nicer to see at the top instead of "jointly readme".
- [x] **Badges** for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
- [ ] **Short description of goals of package,** with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
  - There is no API documentation beyond what is in the vignette in the README. See "must improve" above.
  - There are very helpful descriptive sections of the documentation (e.g., "Usage", "Background") that should be linked here, e.g., in a bulleted list that could help end-users find them faster; e.g., "How to sync sensors"
- [x] Any additional setup required (authentication tokens, etc)
  - Not applicable
- [x] Brief demonstration usage
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
  - Documentation is higher up in the README, which I prefer
- [x] If applicable, how the package compares to other similar packages and/or how it relates to other packages
  - Authors should indicate whether or not this applies.
  - Authors should add a DOI at the top and the example citation at the bottom.

Package structure should follow general community best-practices. In general please consider:

- [x] The documentation is easy to find and understand

  - @xmnlab I have no experience with accelerometers or their data outputs, so I am unable to completely validate the functional claims. However, I have run the vignette and used `ipdb` to interrogate different data structures. Everything looks busy and interesting!
  - No performance claims were made and the software both installs and runs very quickly.
  - **Please add information to the README on how to run the tests.** I have run to Python scripts in the `tests` folder but there is no output so I don't know if they were successful. A standard test suite framework should have some kind of standard output display. It would also be helpful is there was a single test suite that encompassed all tests (so as to avoid running each test suite separately and possible missing one).
- [ ] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.
  - Does not appear to use CI. @xmnlab can you comment on the expectations here? Many CI services are not free, and I would hate for that to be a barrier to any software published in PyOpenSci.
  - This would seem to be the sum of every other effort described here, so I would check this off when everything else is done.

Estimated hours spent reviewing: 1 hour (so far).

---

The software itself appears to be of high quality as it both installs and runs quickly and there is judicious use of modularization (at the development level, not user level), abstraction, and type hinting. The code could benefit from more in-line comments and both class-level and module-level docstings, however.

Please comment on "If applicable, how the package compares to other similar packages and/or how it relates to other packages."
