### 2021-12-09 #45 @xmnlab
@arthur-e thank you for your detailed review. I really appreciate that.

I will check detailed the points you mentioned me and I will give more feedback soon.

About CI, generally, the common CI platforms have a free tier with some limitations, but for small projects that is enough.

Jointly is using github actions and you can check that here https://github.com/hpi-dhc/jointly/actions

And the configuration files are available here https://github.com/hpi-dhc/jointly/tree/master/.github/workflows

Let me know if you need any extra information about that or if you have any comment.

Thank you so much!
