### 2021-12-09 036ce21
Add some fixes for the review (#9)
* Add fixes for the first review
* fix deploy-test
