### 2021-12-09 #45 @enra64
Hey, thank you very much for the review! As @xmnlab already mentioned, we do use CI, and it's with GitHub actions. Additionally, I would like to mention that we have the relevant metadata in `pyproject.toml`, which is kinda the equivalent of `setup.py` for poetry.

As we use the `pytest` framework, tests can be run by calling `py.test` in the repository root. I've added this information to CONTRIBUTING.rst.

It's a good hint that the code reference is broken! I've fixed the generation now.
