### 2021-12-09 #45 @arthur-e
@enra64 Thanks for fixing the API documentation! It looks like there is an object, `jointly.segement_selector`, that may not actually exist in the code base. [It appears in the documentation](https://jointly.readthedocs.io/en/latest/reference.html#jointly-segment-selector), under-specified, because it is written in `docs/reference.rst`.
