### 2021-12-09 #45 @arthur-e
@enra64 Finally, please add information to the README on how you would like people to cite their use of `jointly`. For example, some libraries have DOIs that are automatically generated, e.g., by Zenodo, for specific releases and, in addition to a citation example in the README, the DOI is displayed as a badge at the top of the README.
