+++ README.rst
@@ -29,7 +29,7 @@ Jointly
 .. image:: https://zenodo.org/badge/303936309.svg
    :target: https://zenodo.org/badge/latestdoi/303936309
 
-jointly is a python package for synchronizing sensors with accelerometer data.
+Jointly is a python package for synchronizing sensors with accelerometer data.
 You need this package if you're a researcher who has recorded accelerometer data (plus possibly other data) on multiple sensors and want to precisely synchronize the multiple data streams.
 Specifically, shake all your sensors in a box before and after a trial, and jointly will find these shakes, remove any temporal offset between sensors, and stretch the data so every clock aligns to a reference sensor.
 Jointly ingests and produces ``pandas`` ``DataFrame`` objects.
