+++ CITATION.cff
@@ -16,8 +16,8 @@ authors:
 - family-names: "Arnrich"
   given-names: "Bert"
   orcid: "https://orcid.org/0000-0001-8380-7667"
-title: "jointly - a signal synchronizer"
-version: 2.0.4
-doi: 10.5281/zenodo.1234
-date-released: 2017-12-18
-url: "https://github.com/github/linguist"
+url: "https://github.com/hpi-dhc/jointly"
+doi: "https://zenodo.org/badge/latestdoi/303936309"
+title: "Jointly: A Python package for synchronizing multiple sensors with accelerometer data"
+version: 1.0.3
+date-released: 2021-12-09
