### 2021-12-09 aea4582
Update CITATION.cff
