+++ CITATION.cff
@@ -1,17 +1,17 @@
 cff-version: 1.2.0
 message: "If you use this software, please cite it as below."
 authors:
-- family-names: "Ariane"
-  given-names: "Morassi-Sasso"
+- family-names: "Sasso"
+  given-names: "Ariane"
   orcid: "https://orcid.org/0000-0002-3669-4599"
 - family-names: "Herdick"
   given-names: "Arne"
   orcid: "https://orcid.org/0000-0002-1288-3571"
-- family-names: "Felix"
-  given-names: "Musmann"
+- family-names: "Musmann"
+  given-names: "Felix"
   orcid: "https://orcid.org/0000-0001-5365-0785"
-- family-names: "Justin"
-  given-names: "Albert"
+- family-names: "Albert"
+  given-names: "Justin"
   orcid: "https://orcid.org/0000-0002-6121-792X"
 - family-names: "Arnrich"
   given-names: "Bert"
