+++ README.rst
@@ -156,6 +156,6 @@ This package was created with Cookiecutter_ and the `pyOpenSci/cookiecutter-pyop
 .. _`Full Explanation of the Synchronization`: https://jointly.readthedocs.io/en/latest/background.html#the-syncing-process
 
 Citation
-----------------
+--------
 
 Sasso, A., Herdick, A., Musmann, F., Albert, J., & Arnrich, B. (2021). Jointly: A Python package for synchronizing multiple sensors with accelerometer data (Version 1.0.3) [Computer software]. https://doi.org/https://doi.org/10.5281/zenodo.5770586
