### 2021-12-09 #45 @arianesasso
@enra64 @arthur-e Just added the Zenodo badge + the citation at the README :).
