### 2021-12-26 hpi-dhc/jointly#10 @AlexS12
ModuleNotFoundError when trying to run pytest
* jointly version: master
* Python version: 3.9.7
* Operating System: Ubuntu (inside Windows Subsystem for Linux)

### Description

Trying to test the package while doing the review for PyOpenSci (https://github.com/pyOpenSci/software-review/issues/45) results in different errors

### What I Did

- Create a conda environment: `conda create -n jointly python poetry`
- Clone the project from github.
- `poetry install; pre-commit install`
- Tried to run all tests with `CONTRIBUTING.md` instructions `py.test tests.test_jointly` failed:

```
(jointly) asaez@ap-5cg1140pq3:~/pyopensci/jointly$  py.test tests.test_jointly
================================================= test session starts ==================================================
rootdir: /home/asaez/pyopensci/jointly
collected 0 items

================================================ no tests ran in 0.00s =================================================
ERROR: file or directory not found: tests.test_jointly
```

- Tried to run all tests from `tests` folder and got this error

```
(jointly) asaez@ap-5cg1140pq3:~/pyopensci/jointly/tests$ py.test
================================================= test session starts ==================================================
rootdir: /home/asaez/pyopensci/jointly/tests
collected 5 items / 4 errors / 1 selected

======================================================== ERRORS ========================================================
___________________________________________ ERROR collecting test_helpers.py ___________________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_helpers.py'.
Traceback:
../../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
test_helpers.py:20: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
______________________________________ ERROR collecting test_helpers_plotting.py _______________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_helpers_plotting.py'.
Traceback:
../../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
test_helpers_plotting.py:9: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
_______________________________________ ERROR collecting test_shake_extractor.py _______________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_shake_extractor.py'.
Traceback:
../../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
test_shake_extractor.py:5: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
________________________________________ ERROR collecting test_synchronizer.py _________________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_synchronizer.py'.
Traceback:
../../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
test_synchronizer.py:11: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
=============================================== short test summary info ================================================
ERROR test_helpers.py
ERROR test_helpers_plotting.py
ERROR test_shake_extractor.py
ERROR test_synchronizer.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 4 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
```

- Tried to run the tests from the root folder and got the same error:
```
(jointly) asaez@ap-5cg1140pq3:~/pyopensci/jointly$ pytest
================================================= test session starts ==================================================
rootdir: /home/asaez/pyopensci/jointly
collected 5 items / 4 errors / 1 selected

======================================================== ERRORS ========================================================
________________________________________ ERROR collecting tests/test_helpers.py ________________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_helpers.py'.
Traceback:
../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
tests/test_helpers.py:20: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
___________________________________ ERROR collecting tests/test_helpers_plotting.py ____________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_helpers_plotting.py'.
Traceback:
../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
tests/test_helpers_plotting.py:9: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
____________________________________ ERROR collecting tests/test_shake_extractor.py ____________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_shake_extractor.py'.
Traceback:
../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
tests/test_shake_extractor.py:5: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
_____________________________________ ERROR collecting tests/test_synchronizer.py ______________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_synchronizer.py'.
Traceback:
../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
tests/test_synchronizer.py:11: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
=============================================== short test summary info ================================================
ERROR tests/test_helpers.py
ERROR tests/test_helpers_plotting.py
ERROR tests/test_shake_extractor.py
ERROR tests/test_synchronizer.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 4 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
============================================ 4 errors in 135.08s (0:02:15) =============================================
```
