### 2021-12-26 #45 @AlexS12
Hi there! I was starting my review work today and I am having some problems with running the tests. I have just opened an issue https://github.com/hpi-dhc/jointly/issues/10
