### 2021-12-26 #45 @AlexS12
Hi! Here you can find the first round of my review. It is mainly based on the first contact with the package installation and documentation reading (more to come soon). I have left some suggestions below. Not many blocking aspects so far.

I have enjoyed a lot reviewing your package thanks for the contribution!

  - (Could be improved) Add a link to `CONTRIBUTING.rst` for development installation.
  - (Could be improved) Some plots can be added to show the not-synched vs the synched data, peak identification...
- [ ] **Function Documentation:** for all user-facing functions
  - (Must be improved) `jointly.helpers.get_segment_data`: documentation is missing.
  - (Must be improved) `jointly.helpers.get_stretch_factor`: only looking at the documentation, it is not clear for me what should be passed as a `segment` and `timeshift`. Doc could be improved or an example might be necessary.
  - (Could be improved) `jointly.helpers.normalize`: normalization formula might be added to doc.
  - (Must be improved) `jointly.segment_selector`is empty.
- [X] **Examples** for all user-facing functions
  - (Could be improved) It might be interesting to show `get_sync_params` information in the usage example.
  - (Could be improved) Some graphical examples showing successfull and unsuccessfull `ShakeExtractor` configurations might be useful to understand how the sakes are identified ie. a wrong configuration that might throw a `ShakeMissingException`.
  - (Could be improved) Examples are missing for:
    - `get_equidistant_signals`
    - `get_max_ref_frequency`
    - `get_segment_data`
    - `get_stretch_factor`

   - https://jointly.readthedocs.io/en/latest/contributing.html#get-started Good guide! I find that commenting some lines with a small description of what each command is doing might be useful for newcomers.
   - Some projects prefer developers to submit an issue with the new feature suggestion before pull requests with new implementations. If that is your case I would suggest to add an indication.
   - History section is good! I find it more useful when the pull request link is found next to the new features/bugs corrections when applicable. Might be a good idea to incorporate that to the contributing guidelines.
   - Nice to see your code of conduct!! =)
- [X] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [X] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
- [X] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [X] Any additional setup required (authentication tokens, etc)
    - Not applicable.
- [X] Brief demonstration usage
- [X] Direction to more detailed documentation (e.g. your documentation files or website).
- [X] If applicable, how the package compares to other similar packages and/or how it relates to other packages
    - Not applicable.

Package structure should follow general community best-practices. In general please consider:

- [X] The documentation is easy to find and understand
  - See Documentation section in this review.

  - Problems found to run the tests locally as described in contributing section: https://github.com/hpi-dhc/jointly/issues/10
  - Pending
  - Pending
  - See https://github.com/hpi-dhc/jointly/issues/10
- [X] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

- First round: 2 hours
