### 2021-12-27 353432c
fix some documentation issues
