### 2021-12-27 #45 @enra64
Hey @AlexS12, thank you very much for the very detailed review! Regarding the documentation: some of those functions are expected to be used only internally, so the documentation or parameters might seem a bit weird. I've added this information to `get_stretch_factor` as that might seem especially weird. The test-running examples have been updated, it would be nice if you could test again using conda as I don't have that installed :sweat_smile:

The changes will be merged via PR https://github.com/hpi-dhc/jointly/pull/12
