### 2021-12-27 ce3824b
Fix more documentation issues (#12)
* fix some documentation issues
* update history, version bump
