+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "jointly"
-version = "1.0.3"
+version = "1.0.4"
 description = "Synchronize sensor data from accelerometer shakes"
 authors = [
     "Ariane Morassi Sasso <ariane.morassi-sasso@hpi.de>",
