### 2021-12-27 hpi-dhc/jointly#10 @AlexS12
Hi @enra64! I  am still experimenting this issue. I am confused about the results.

I will investigate/get some advice to see what's going on. If I cannot get an answer and you confirm that you can run the tests locally I will assume this is a configuration problem in my environment for the review. However, as someone else might hit this same issue, it would be great to have an answer.

```
(jointly) asaez@ap-5cg1140pq3:~/pyopensci/jointly$ pytest tests
================================================= test session starts ==================================================
rootdir: /home/asaez/pyopensci/jointly
plugins: anyio-3.4.0
collected 5 items / 4 errors / 1 selected

======================================================== ERRORS ========================================================
________________________________________ ERROR collecting tests/test_helpers.py ________________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_helpers.py'.
Traceback:
../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
tests/test_helpers.py:20: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
___________________________________ ERROR collecting tests/test_helpers_plotting.py ____________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_helpers_plotting.py'.
Traceback:
../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
tests/test_helpers_plotting.py:9: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
____________________________________ ERROR collecting tests/test_shake_extractor.py ____________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_shake_extractor.py'.
Traceback:
../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
tests/test_shake_extractor.py:5: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
_____________________________________ ERROR collecting tests/test_synchronizer.py ______________________________________
ImportError while importing test module '/home/asaez/pyopensci/jointly/tests/test_synchronizer.py'.
Traceback:
../../anaconda3/envs/jointly/lib/python3.9/importlib/__init__.py:127: in import_module
tests/test_synchronizer.py:11: in <module>
    from tests.parquet_reader import get_parquet_test_data
E   ModuleNotFoundError: No module named 'tests.parquet_reader'
=============================================== short test summary info ================================================
ERROR tests/test_helpers.py
ERROR tests/test_helpers_plotting.py
ERROR tests/test_shake_extractor.py
ERROR tests/test_synchronizer.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 4 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
============================================ 4 errors in 133.77s (0:02:13) =============================================
```

Same error obtained with:

```
(jointly) asaez@ap-5cg1140pq3:~/pyopensci/jointly$  python -m pytest tests
(jointly) asaez@ap-5cg1140pq3:~/pyopensci/jointly/tests$  python -m pytest
(jointly) asaez@ap-5cg1140pq3:~/pyopensci/jointly$ poetry run python -m pytest
```
Last one should be equivalent to CI config file:
```
      - name: Test with pytest
        run: |
          poetry run python -m pytest
```

The solution for me is to add an `__init__.py` inside tests folder.

```
(jointly) asaez@ap-5cg1140pq3:~/pyopensci/jointly$ touch tests/__init__.py
(jointly) asaez@ap-5cg1140pq3:~/pyopensci/jointly$ pytest tests
================================================= test session starts ==================================================
rootdir: /home/asaez/pyopensci/jointly
plugins: anyio-3.4.0
collected 27 items

tests/test_examples.py .....                                                                                     [ 18%]
tests/test_helpers.py .........                                                                                  [ 51%]
tests/test_helpers_plotting.py ..                                                                                [ 59%]
tests/test_shake_extractor.py ...                                                                                [ 70%]
tests/test_synchronizer.py ........                                                                              [100%]

============================================ 27 passed in 134.55s (0:02:14) ============================================
```

However, I cannot understand why I do not obtain the same results in as in your local envioronment or the Github CI (https://github.com/hpi-dhc/jointly/runs/4642228510?check_suite_focus=true)
```
Run poetry run python -m pytest
  poetry run python -m pytest
  shell: /usr/bin/bash -e {0}
  env:
    pythonLocation: /opt/hostedtoolcache/Python/3.9.9/x64
    LD_LIBRARY_PATH: /opt/hostedtoolcache/Python/3.9.9/x64/lib
    VENV: .venv/bin/activate
rootdir: /home/runner/work/jointly/jointly
collected 27 items

tests/test_examples.py .....                                             [ 18%]
tests/test_helpers.py .........                                          [ 51%]
tests/test_helpers_plotting.py ..                                        [ 59%]
tests/test_shake_extractor.py ...                                        [ 70%]
tests/test_synchronizer.py ........                                      [100%]

============================== 27 passed in 7.46s ==============================
```
