### 2021-12-27 hpi-dhc/jointly#10 @astrojuanlu
I can't reproduce with venv + poetry, I get the test suite passing normally. conda + poetry is showing some weird behavior on my computer...

```
$ which poetry
/home/juanlu/.mambaforge/envs/jointly/bin/poetry
$ poetry install
...
  • Updating alabaster (0.7.12 /home/juanlu/.local/lib/python3.9/site-packages -> 0.7.12)  # ????????????????
```

(not a fan of Poetry in general, so I won't spend more resources debugging this)

@AlexS12 can you try removing your `__init__.py` workaround and run the following?

```
>>> import tests
>>> tests
???
>>> tests.__file__
???
>>> import tests.parquet_reader
>>> tests.parquet_reader.__file__
???
```

I suspect you might have a `tests` package somewhere else that produces the error you see. This is what I get:

```
>>> import tests
>>> tests.__file__
>>> tests  # "namespace package" as expected since there is no __init__.py, see PEP 420
<module 'tests' (namespace)>
>>> import tests.parquet_reader
>>> tests.parquet_reader.__file__
'/home/juanlu/Projects/Other/tmp/jointly/tests/parquet_reader.py'
```

However, if for any reason there is a `tests` directory somewhere else in your `sys.path`, it will get imported first. For example:

```
(.venv) $ mkdir /tmp/tests
(.venv) $ touch /tmp/tests/__init__.py
(.venv) $ export PYTHONPATH=/tmp:$PYTHONPATH
(.venv) $ python
Python 3.8.10 (default, Nov 26 2021, 20:14:08)
[GCC 9.3.0] on linux
>>> import tests
>>> tests.__file__  # Woooops
'/tmp/tests/__init__.py'
>>> tests
<module 'tests' from '/tmp/tests/__init__.py'>
>>> import tests.parquet_reader  # This will now fail!
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ModuleNotFoundError: No module named 'tests.parquet_reader'
```
