### 2021-12-27 hpi-dhc/jointly#10 @AlexS12
@astrojuanlu, you hit the nail with that! Thanks!

```
Python 3.9.7 (default, Sep 16 2021, 13:09:58)
[GCC 7.5.0] :: Anaconda, Inc. on linux
>>> import tests
>>> tests.__file__
'/home/asaez/anaconda3/envs/jointly/lib/python3.9/site-packages/tests/__init__.py'
>>> dir(tests)
['__builtins__', '__cached__', '__doc__', '__file__', '__loader__', '__name__', '__package__', '__path__', '__spec__']
>>> import tests.parquet_reader.__file__
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ModuleNotFoundError: No module named 'tests.parquet_reader'
```

Given the content of `anaconda3/envs/jointly/lib/python3.9/site-packages/tests/conftest.py`, it seems to come from [pastel](https://pypi.org/project/pastel/) package:

```
# -*- coding: utf-8 -*-

import pytest

from pastel import Pastel
from pastel.style import Style
from pastel.stack import StyleStack

@pytest.fixture
def pastel():
    return Pastel(True)

@pytest.fixture
def non_decorated_pastel():
    return Pastel(False)

@pytest.fixture
def style():
    return Style()

@pytest.fixture
def stack():
    return StyleStack()
```
I don't know where the pastel dependency comes from. Given your results, the problem seems to be caused by the way conda installs pastel, but it might be a good idea to avoid the collision somehow. If the `__init__.py` is not a good idea, what would it be? maybe a `conftest.py`? Otherwise, it can be noted in the instructions that tests won't run in a conda environment.
