### 2021-12-27 hpi-dhc/jointly#10 @astrojuanlu
There's two parts of this issue:

1. https://github.com/sdispater/pastel/issues/7 (almost two years old... unlikely to be fixed?)
2. Tests of jointly doing `from tests.xxx import yyy` (not a big deal, but pytest encourages using fixtures for that, and other packages like Astropy ship the test utilities)

In any case, I'd say that this is mainly a problem with Pastel. I'm not sure if it's a transitive dependency of jointly?
