### 2021-12-27 hpi-dhc/jointly#10 @AlexS12
Apparently, pastel is installed when installing poetry with conda:

```
(base) asaez@ap-5cg1140pq3:~/pyopensci/jointly$ conda create -n jointly-pastel python poetry
Collecting package metadata (current_repodata.json): done
Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
Solving environment: done

## Package Plan ##

  environment location: /home/asaez/anaconda3/envs/jointly-pastel

  added / updated specs:
    - poetry
    - python

The following NEW packages will be INSTALLED:

  ...
  pastel             pkgs/main/noarch::pastel-0.2.1-py_0
  ...
```
