### 2021-12-27 hpi-dhc/jointly#10 @astrojuanlu
From what I understand, pastel will no longer be a transitive dependency of Poetry very soon:

https://github.com/python-poetry/poetry/pull/3618

(the reason is that the new Cleo removed its dependency on Clikit https://github.com/sdispater/cleo/pull/93, which is the one that depended on pastel https://github.com/sdispater/clikit/blob/1da34f16c50a2a7d4a22d21250cb6268c7b3eefd/pyproject.toml#L20)

Because of the temporary nature of the problem, I'd rather not give it more importance 😅
