### 2021-12-29 #45 @xmnlab
----------------------------------------------
🎉 Jointly has been approved by pyOpenSci! Thank you @enra64 for submitting Jointly and many thanks to @AlexS12 and @arthur-e  for reviewing this package! 😸

- [ ] Activate Zenodo watching the repo if you haven't already done so.
- [ ] Add the badge for pyOpenSci peer-review to the README.md of Jointly. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/45)`
- [ ] Add Jointly to the pyOpenSci website. @enra64 , please open a pr to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/packages.yml): to add your package and name to the list of contributors
- [ ] @enra64 @arthur-e @AlexS12,  if you have time and are open to being listed on our website, please add yourselves to [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/contributors.yml) via a pr so we can list you on our website as contributors!

All -- if you have any feedback for us about the review process please feel free to share it here. We are always looking to improve our process and our documentation in the [contributing-guide](https://www.pyopensci.org/contributing-guide). We have also been updating our documentation to improve the process so all feedback is appreciated!

@lwasser let me know if there is anything else missing from the editor side :)
thank you for all the guidance :)

Thanks everyone! it was a pleasure to be acting here as editor
I hope you all have a great new year :tada: :tada: :tada:
