### 2022-01-10 ac4b65e
Update README.rst
Added pyOpenSci badge and new Zenodo DOI
