### 2022-01-10 #45 @arianesasso
Dear @xmnlab, @arthur-e, @AlexS12, and @lwasser, thank you so much for your great work and help!

I hope you had a good start of the year!

I have activated the Zenodo watch, created the release, added the badge, and created the PR :). This review process is awesome and I am looking forward to contributing again 🙌
