### 2022-01-11 60653e6
Update CONTRIBUTORS.rst
