+++ CONTRIBUTORS.rst
@@ -7,10 +7,16 @@ Contributors
 
 * Felix Musmann: <felix.musmann@student.hpi.de>
 * Arne Herdick <arne.herdick@student.hpi.de>
-* Ariane Morassi-Sasso <ariane.morassi-sasso@hpi.de>
+* Ariane Morassi Sasso <ariane.morassi-sasso@hpi.de>
 
 
 Bug Fixes and Error Spotting
 ----------------------------
 
 * Justin Albert <justin.albert@hpi.de>
+
+
+Mentor
+----------------------------
+
+* Prof. Dr. Bert Arnrich <bert.arnrich@hpi.de>
