### 2022-01-11 b56fca2
Update CONTRIBUTORS.rst
