### 2022-01-12 #45 @xmnlab
> @xmnlab You forgot to hit the `Close` button!

@astrojuanlu thanks for catching that. I was very sick yesterday and I am still recovering :) so I am a bit slow today XD

@NickleDave thanks for closing the issue :)
