### 2022-09-28 #45 @lwasser
Good morning @AlexS12 and @enra64 ! just a friendly reminder to take 5-10 minutes to fill out our survey . We really appreciate it. Arne if you can have all of the maintainers of your package fill it out that would be great. We will stay in touch with package authors - checking in 1-2 times a year to ensure the package continues to be maintained and see better understand challenges that you face. Thank you in advance for helping us by filling out the survey!! 🙌

[🔗 HERE IS THE SURVEY LINK 🔗 ](https://forms.gle/ZZXcD1hr18tYmY2p6)
